+++ static/CHANGELOG.md
@@ -1,3 +1,4 @@
+- transfer /formatt --> /format (ysturasp, 2025-12-11) [24 files changed, 375 insertions(+), 1783 deletions(-)]
 - add badge message in toilets page (ysturasp, 2025-12-10) [3 files changed, 10 insertions(+)]
 - refactor toilet location handling and update UI for improve user experience in ToiletResult component (ysturasp, 2025-12-10) [6 files changed, 151 insertions(+), 78 deletions(-)]
 - remove manual login functionality and associated API endpoint for test user authentication (ysturasp, 2025-12-10) [5 files changed, 3 insertions(+), 178 deletions(-)]